move markdown generation to own file
